[2126] Adds KDocs and updates params for new composable function
