+++ livestream-compose-sample/src/main/java/io/getstream/livestream/compose/models/LiveStreamType.kt
@@ -1,4 +1,4 @@
-package io.getstream.livestream.compose
+package io.getstream.livestream.compose.models
 
 enum class LiveStreamType {
     Youtube,
