+++ livestream-compose-sample/src/main/java/io/getstream/livestream/compose/streams/CameraLiveStream.kt
@@ -7,39 +7,24 @@ import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.camera.view.PreviewView
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.LifecycleOwner
-import io.getstream.chat.android.compose.handlers.SystemBackPressedHandler
-import io.getstream.chat.android.compose.ui.messages.list.MessageList
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.viewmodel.messages.MessageComposerViewModel
 import io.getstream.chat.android.compose.viewmodel.messages.MessageListViewModel
-import io.getstream.livestream.compose.LiveStreamHeader
-import io.getstream.livestream.compose.LiveStreamMessage
-import io.getstream.livestream.compose.LiveStreamMessageList
-import io.getstream.livestream.compose.LivestreamComposer
+import io.getstream.livestream.compose.ui.CommentsBox
+import io.getstream.livestream.compose.ui.LiveStreamHeader
 
 /**
  * Shows a camera preview surface in a view component that relies on [MessageListViewModel]
@@ -83,7 +68,7 @@ fun CameraLiveStream(
                 },
                 modifier = Modifier.fillMaxSize(),
             )
-            Box(
+            CommentsBox(
                 modifier = Modifier
                     .background(
                         brush = Brush.verticalGradient(
@@ -92,14 +77,10 @@ fun CameraLiveStream(
                             1050f,
                         )
                     )
-                    .align(Alignment.BottomCenter)
-            ) {
-                Column {
-                    LiveStreamMessageList(listViewModel)
-                    LivestreamComposer(composerViewModel)
-                }
-            }
-
+                    .align(Alignment.BottomCenter),
+                composerViewModel,
+                listViewModel
+            )
             LiveStreamHeader {
                 onBackPressed()
             }
