+++ livestream-compose-sample/src/main/java/io/getstream/livestream/compose/streams/YoutubeLiveStream.kt
@@ -2,7 +2,6 @@ package io.getstream.livestream.compose.streams
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -13,16 +12,16 @@ import androidx.compose.ui.graphics.Color
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.viewmodel.messages.MessageComposerViewModel
 import io.getstream.chat.android.compose.viewmodel.messages.MessageListViewModel
-import io.getstream.livestream.compose.LiveStreamHeader
-import io.getstream.livestream.compose.LiveStreamMessageList
-import io.getstream.livestream.compose.LivestreamComposer
 import io.getstream.livestream.compose.players.YoutubePlayer
+import io.getstream.livestream.compose.ui.CommentsBox
+import io.getstream.livestream.compose.ui.LiveStreamHeader
 
 /**
  * Shows a Youtube view component that relies on [MessageListViewModel]
  * and [MessageComposerViewModel] to connect all the chat data handling operations.
  *
  * @param modifier - Modifier for styling.
+ * @param videoId - String id of Youtube video url
  * @param composerViewModel - [MessageComposerViewModel] for manging message input field
  * @param listViewModel - [MessageListViewModel] The ViewModel that stores all the data and
  * business logic required to show a list of messages. The user has to provide one in this case,
@@ -32,6 +31,7 @@ import io.getstream.livestream.compose.players.YoutubePlayer
 @Composable
 fun YoutubeLiveStream(
     modifier: Modifier = Modifier,
+    videoId: String,
     composerViewModel: MessageComposerViewModel,
     listViewModel: MessageListViewModel,
     onBackPressed: () -> Unit
@@ -42,8 +42,10 @@ fun YoutubeLiveStream(
 
     ChatTheme {
         Box(modifier = modifier.fillMaxSize()) {
-            YoutubePlayer()
-            Box(
+            YoutubePlayer(
+                videoId = videoId
+            )
+            CommentsBox(
                 modifier = Modifier
                     .background(
                         brush = Brush.verticalGradient(
@@ -52,13 +54,10 @@ fun YoutubeLiveStream(
                             1050f,
                         )
                     )
-                    .align(Alignment.BottomCenter)
-            ) {
-                Column {
-                    LiveStreamMessageList(listViewModel)
-                    LivestreamComposer(composerViewModel)
-                }
-            }
+                    .align(Alignment.BottomCenter),
+                composerViewModel,
+                listViewModel
+            )
             LiveStreamHeader {
                 onBackPressed()
             }
