+++ livestream-compose-sample/src/main/java/io/getstream/livestream/compose/ui/LiveStreamHeader.kt
@@ -1,4 +1,4 @@
-package io.getstream.livestream.compose
+package io.getstream.livestream.compose.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -13,9 +13,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
-
+import io.getstream.livestream.compose.R
 
 /**
  * View component to add support for header and a back button icon
@@ -41,7 +42,7 @@ fun LiveStreamHeader(
         ) {
             Icon(
                 imageVector = Icons.Default.ArrowBack,
-                contentDescription = null,
+                contentDescription = stringResource(id = R.string.accessibilityBackButton),
                 tint = ChatTheme.colors.textHighEmphasis,
             )
         }
