+++ livestream-compose-sample/src/main/res/values/strings.xml
@@ -8,4 +8,10 @@
     <string name="desc3">Learn how to combine approaches with declarative API for native management and deployment across clusters.</string>
     <string name="desc4">Data management &amp; Analytics best practices - Driving data literacy</string>
 
+    <string name="accessibilityBackButton">Go back</string>
+    <string name="accessibilityArtworkImage">Artwork image view</string>
+    <string name="accessibilityIcon">Icon</string>
+    <string name="accessibilitySwitchTheme">Switch theme</string>
+    <string name="accessibilityGridToggle">Grid toggle</string>
+
 </resources>
